Append suffix to role names for schemas that must be created in multiple DBs

+++ shared/schema/drop_role.sql
@@ -1,2 +1,2 @@
-DROP ROLE IF EXISTS ${dbSchema}_read;
-DROP ROLE IF EXISTS ${dbSchema}_write;
+DROP ROLE IF EXISTS ${dbSchema}${roleSuffix}_read;
+DROP ROLE IF EXISTS ${dbSchema}${roleSuffix}_write;
